Refactor server utils path within source code
In this change, we've updated the import path of the server utilities across multiple files. Instead of importing the utilities from "source/server/admin/utils," we are now importing them from "source/server/utils". This restructuring has been done to consolidate the utilities under "source/server/utils" instead of having them under the "admin" sub-directory. This is a part of a larger codebase refactoring process to improve the file structure and readability.